Move generator test outside of `__tests__` folders
No need to have a separate folder to hold an unique file.
