+++ .eslintrc.json
@@ -7,6 +7,15 @@
     "parserOptions": {
         "project": "./tsconfig.json"
     },
+    "rules": {
+        "import/no-extraneous-dependencies": ["error", {
+           "devDependencies": [
+                "**/__tests__/**",
+                "**/*.test.ts"
+            ],
+            "optionalDependencies": false
+        }]
+    },
     "overrides": [
         {
             "files": ["**/__tests__/*.ts", "*.test.ts"],
