+++ generators/express/index.test.ts
@@ -8,12 +8,12 @@ describe('When running the generator', () => {
     let root: string;
 
     beforeAll(async () => {
-        const result = await helpers.run(path.resolve(__dirname, '../../root'))
+        const result = await helpers.run(path.resolve(__dirname, '../root'))
             .withPrompts({ contactEmail: 'test@example.com' });
 
         root = result.cwd;
 
-        await helpers.run(path.resolve(__dirname, '..'))
+        await helpers.run(__dirname)
             .cd(root)
             .withOptions({ skipInstall: false })
             .withArguments(['test']);
@@ -27,9 +27,19 @@ describe('When running the generator', () => {
         await execa('yarn', ['build'], { cwd: path.resolve(root, 'test') });
     });
 
+    test('It generates a project which correctly lints', async () => {
+        await execa('yarn', ['lint'], { cwd: path.resolve(root, 'test') });
+    });
+
     test('It generates a docker-compose.yaml with a version fields', async () => {
         const all = YAML.parse(await fs.promises.readFile(path.resolve(root, 'docker-compose.yaml'), 'utf8'));
 
         expect(all.version).toBeDefined();
+    });
+
+    test('It extends the ansible configuration', async () => {
+        const all = YAML.parse(await fs.promises.readFile(path.resolve(root, 'ansible/group_vars/all.yaml'), 'utf8'));
+
+        expect(all.test_env).toBeDefined();
     });
 });
