+++ tsconfig.build.json
@@ -2,6 +2,7 @@
     "extends": "./tsconfig.json",
     "exclude": [
         "generators/*/templates",
-        "**/__tests__"
+        "**/__tests__",
+        "**/*.test.ts"
     ]
 }
